Update README.md
The current pace of technological advancements has the most profound impact on enabling how chemical manufacturers transform themselves to respond to market trends and deliver an entirely new customer experience. Intelligent technologies promise to bring great benefits such as productivity and efficiency gains, enabling innovative new business models and generating new revenue streams. Following intelligent technologies are instrumental in helping chemical companies respond to market trends.
Blockchain
A blockchain supply chain can help participants record price, date, location, quality, certification, and other relevant information to effectively manage the supply chain. The availability of this information within blockchain can increase traceability of material supply chain, lower losses from counterfeit and grey markets, improve visibility and compliance over outsourced contract manufacturing, and potentially enhance an organization's position as a leader in responsible manufacturing. The blockchain model of trust, through massively distributed digital consensus, could reshape supply chains and commerce across the entire digital economy. Digitalizing the bill of lading document as part of the international ocean shipping process can be used as a good example.
Conversational AI
Advances in machine learning are enabling algorithms to become highly accurate in natural-language understanding and in image and speech recognition, which is especially useful in after-service and call-centre activities. Voice interface will be the go-to technology for the next generation of applications, allowing room for greater simplicity, mobility, and efficiency while increasing worker productivity and reducing the need for training.
Data Platform to Manage Experience
Leaders are connecting operational data from companies’ business systems (what is happening) with the experience data coming from customers and employees (why is it happening) to get 360-degree views and actionable insights, so they can deliver an overall enhanced experience.
It is expected that by the end of the fiscal Year 2022-23, around a 57% increase on average will happen, in the contribution of machines and algorithms to be specific.
Conclusion
Concluding the series, the chemical industry and the technology are evolving at a very high pace, driven by the industry’s mega-trends in terms of strategy, priorities and technologies. To keep up with the market trends and to stay as a leader in the chemical industry it is vital for chemical organizations and their leaders to stay informed and continuously evolve themselves, using better processes and practices. Subscribing to our thought leadership articles is a great way to ensure that you have an edge over the competition, derived with updates about what is best in the market with detailed insights about the latest trends in the industry. 
•	Solulever's focus on digital transformation and their Brabo Edge platform for manufacturing connectivity and intelligence seems to align with the growing need for technology-driven solutions in today's business landscape.By leveraging the capabilities of the Brabo Edge platform, businesses can likely streamline their manufacturing processes, improve efficiency, and gain valuable insights into their operations. Digital transformation has become crucial for enterprises to stay competitive and adapt to the evolving market demands.
For full article please click the below link:
https://www.linkedin.com/feed/update/urn:li:activity:7054680089170518016/